src/proxy/requirements.txt
sqlglot==10.6.2
psycopg2-binary==2.9.6
python-dotenv==0.19.2
pytds==1.11.0
uvicorn==0.18.2
fastapi==0.78.0
